added another option to the No Receiver warning panel.
Now the user is presented with an option to create a
simple Socket Receiver on one of the standard ports.

This is the most likely option for new Chainsaw users,
and is a more user friendly option then having to specify
it via configs, or even create one via the GUI.

Very handy for testing purposes too.