A large number of changes because of the move to the new Rule structure.
Some code has been deprecated as an effort to tag which
code needs either reworking or replacement because of this change.

The Logger Tree panel now has a working implementation of the
"Focus On" principle, allowing temporary/quick view filtering
based on the focused on Logger and it's children.

The other filtering mechanisms have not been hooked up at this time.

Colouring is still not in operation.