VFSPlugin now looking a bit better, although still no actual functionality yet
This plugin automatically adds a 'local' VFS file system node to the tree just for testing purposes.

Eventually one will be able to browse the file system root to find the children, and navigate
a directory hierachy, and then select a file to view in the detail table.

I will also add a 'preview' panel below the table which, hopefully, will be able to display
the first X lines of text in the file.