Commiting patch contributed by Stephen Pain that fixes the patch that I committed last week to fix a problem Stephen found originally.....
Maybe I should have just let Stephen do it in the first place... :)

It sucks that this won't support unlimited XML size for an event, but it is better than taking forever to load a file, or causing an error.

My only change is that I append a "\n" after every line so the stack traces appear nicely inside Chainsaw.