First part of moving out Chainsaw into it's own module.  This tidys up the log4j build process to not need to reference chainsaw at all.
Note: I've explicity added the ListModelAppender into the log4j core via this change, while
this is an Appender that Chainsaw uses (and maybe it won't forever), it is a generally useful feature outside the scope of Chainsaw.   I'm happy to move this into the Chainsaw module if
no-one thinks it's of much use in the core.